+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Fixed
+
+* Retry if downloading of instruments has error message "Unavailable" (#139)
+
 ## 0.7.9 &ndash; 2021-12-08
 
 ## 0.7.8 &ndash; 2021-11-02